refactor copy worker
