+++ desktop/src/main/fileProcessing/actions/getFolderBuffer.ts
@@ -4,7 +4,6 @@ import { app } from "electron";
 
 type WorkerData = {
   source: string;
-  destination?: string;
   batchSize?: number;
 };
 
@@ -20,7 +19,6 @@ type FileBufferObj = {
  * @param workerPool - The WorkerPool instance to manage worker threads.
  * @param filePath - The source folder path to be processed.
  * @param isDev - Is running in the development build (npm run dev).
- * @param storeFile - If the file should be stored in the app data.
  * @param onProgress - Callback for progress updates.
  * @param onCompletion - Callback for final completion data.
  * @returns A Promise that resolves when the worker processes complete.
@@ -29,7 +27,6 @@ export const getFolderBuffer = async (
   pool: WorkerPool,
   filePath: string,
   isDev: boolean,
-  storeFile: boolean,
   onProgress?: (data: { progressPercentage: number; source: string }) => void,
   onMissingPath?: (data: { path: string }) => void,
   onEmptyFolder?: (data: { path: string }) => void,
@@ -43,15 +40,9 @@ export const getFolderBuffer = async (
     ? path.resolve(__dirname, "../es-workers/copyWorker.js")
     : path.join(app.getAppPath(), "../../resources/copyWorker.cjs");
 
-  const destinationPath = isDev
-    ? path.resolve(__dirname, "../../resources/file-list")
-    : path.join(app.getAppPath(), "../../resources/file-list");
-
   const workerData: WorkerData = {
     source: filePath,
   };
-
-  if (storeFile) workerData.destination = destinationPath;
 
   try {
     pool.on("progress", (data) => {
@@ -72,7 +63,7 @@ export const getFolderBuffer = async (
 
     await pool.runTask(workerScript, workerData);
   } catch (error) {
-    console.error(`[Action] Failed to process folder ${filePath}:`, error);
+    console.error(`[Copy Action] Failed to process folder ${filePath}:`, error);
     if (onCompletion) onCompletion({ success: false, error });
   }
 };
