+++ desktop/src/main/fileProcessing/actions/getFolderMetadata.ts
@@ -4,7 +4,6 @@ import { app } from "electron";
 
 type WorkerData = {
   source: string;
-  destination?: string;
   batchSize?: number;
 };
 
@@ -14,7 +13,6 @@ type WorkerData = {
  * @param workerPool - The WorkerPool instance to manage worker threads.
  * @param filePath - The source folder path to be processed.
  * @param isDev - Is running in the development build (npm run dev).
- * @param storeFile - If the file should be stored in the app data.
  * @param onProgress - Callback for progress updates.
  * @param onCompletion - Callback for final completion data.
  * @returns A Promise that resolves when the worker processes complete.
@@ -23,7 +21,6 @@ export const getFolderMetadata = async (
   pool: WorkerPool,
   filePath: string,
   isDev: boolean,
-  storeFile: boolean,
   onProgress?: (data: { progressPercentage: number; source: string }) => void,
   onMissingPath?: (data: { path: string }) => void,
   onEmptyFolder?: (data: { path: string }) => void,
@@ -39,19 +36,9 @@ export const getFolderMetadata = async (
     ? path.resolve(__dirname, "../es-workers/metadataWorker.js")
     : path.join(app.getAppPath(), "../../resources/metadataWorker.cjs");
 
-  const destinationPath = isDev
-    ? path.resolve(__dirname, "../../resources/file-list")
-    : path.join(app.getAppPath(), "../../resources/file-list");
-
   const metadataWorkerData: WorkerData = {
     source: filePath,
   };
-
-  if (storeFile)
-    metadataWorkerData.destination = path.join(
-      destinationPath,
-      "/metadata.json"
-    );
 
   try {
     pool.on("progress", (data) => {
