+++ desktop/src/main/fileProcessing/workers/utilities/index.ts
@@ -1,8 +1,8 @@
 export * from "./calculateChecksum";
 export * from "./countFiles";
 export * from "./doesDirectoryExist";
-export * from "./ensureDirectoryExists";
 export * from "./formatFileSize";
 export * from "./generateExtendedMetadata";
 export * from "./generateMetadataInBatches";
+export * from "./generateBuffersInBatches";
 export * from "./writeOrAppendMetadata";
